HECK YES
Using the TextLoad function, the JSON is read in, parsed, parsed, parsed, and printed to the button interface. It is now possible to tell the files apart.
Updates will be necessary to update the buttons on deletion of the file, and to indicate a lack of a file, in doing so deactivating the interactibility of the button.